chore: remove unused functions
